Password input refactored to component
