+++ symfony/client/src/components.ts
@@ -28,6 +28,7 @@ import Form from '@orangehrm/oxd/src/core/components/Form/Form.vue';
 import FormRow from '@orangehrm/oxd/src/core/components/Form/FormRow.vue';
 import FormActions from '@orangehrm/oxd/src/core/components/Form/FormActions.vue';
 import InputField from '@orangehrm/oxd/src/core/components/InputField/InputField.vue';
+import InputGroup from '@orangehrm/oxd/src/core/components/InputField/InputGroup.vue';
 import TableFilter from '@orangehrm/oxd/core/components/TableFilter/TableFilter.vue';
 import Grid from '@orangehrm/oxd/core/components/Grid/Grid.vue';
 import GridItem from '@orangehrm/oxd/core/components/Grid/GridItem.vue';
@@ -47,6 +48,7 @@ export default {
     app.component('oxd-form-row', FormRow);
     app.component('oxd-form-actions', FormActions);
     app.component('oxd-input-field', InputField);
+    app.component('oxd-input-group', InputGroup);
     app.component('oxd-grid', Grid);
     app.component('oxd-grid-item', GridItem);
     app.component('oxd-table-filter', TableFilter);
